refactor: move investigador components
