+++ src/app/ui/cards/index.ts
@@ -1,2 +1,4 @@
-export * from "./investigador"
+export * from "./CardModal"
+export * from "./CardSkeleton"
+export * from "./HorizontalCard"
 export * from "./proyecto"
