+++ src/app/ui/form/create/index.ts
@@ -1,2 +1,1 @@
-export * from "./InvestigadorCreateForm"
 export * from "./ProyectoCreateForm"
