+++ src/app/ui/menu/useAddParticipante.ts
@@ -1,12 +1,13 @@
 import { useEffect, useState } from "react"
 import { useDebouncedCallback } from "use-debounce"
 
-import type { InvestigadorMinimumDataType, ParticipaType } from "@/app/utils"
+import type { ParticipaType } from "@/app/utils"
 import {
   fetchInvestigadoresByQuery,
   fetchInvestigadoresNotParticipatingInProject,
   fetchNombresDeAutor,
 } from "@/db"
+import type { InvestigadorMinimumDataType } from "@/investigadores"
 
 export const useAddParticipante = ({
   codigo,
