--- src/app/utils/investigador/index.ts
@@ -1,1 +1,0 @@
-export * from "./add"
