+++ src/app/utils/types/items.ts
@@ -1,18 +1,3 @@
-export type InvestigadorType = {
-  email: string
-  nombre: string
-  apellidos: string
-  universidad: string
-  departamento: string
-  area: string
-  figura: string
-}
-
-export type InvestigadorMinimumDataType = Pick<
-  InvestigadorType,
-  "email" | "nombre" | "apellidos"
->
-
 export type ProyectoType = {
   codigo: string
   ip: string
