+++ src/app/utils/validation/index.ts
@@ -1,3 +1,2 @@
 export * from "./constants"
-export * from "./investigador"
 export * from "./project"
