+++ src/investigadores/card/index.ts
