+++ src/investigadores/form/index.ts
@@ -1,0 +1,2 @@
+export * from "./create/InvestigadorCreateForm"
+export * from "./edit/EditInvestigadorForm"
