+++ src/investigadores/index.ts
@@ -1,0 +1,3 @@
+export * from "./card"
+export * from "./form"
+export * from "./types"
