+++ src/investigadores/utils/add.ts
@@ -1,12 +1,9 @@
 "use server"
 
-import type { InvestigadorType } from "@/db"
 import { createInvestigador } from "@/db"
 
-import {
-  getInvestigadorErrors,
-  validateInvestigadorErrors,
-} from "../validation/investigador"
+import type { InvestigadorType } from "../types"
+import { getInvestigadorErrors, validateInvestigadorErrors } from "./validation"
 
 export const addInvestigador = async (investigador: InvestigadorType) => {
   if (validateInvestigadorErrors(getInvestigadorErrors(investigador))) {
