+++ src/investigadores/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./add"
+export * from "./validation"
